server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pyyaml>=6.0.1
jinja2>=3.1.3
python-multipart>=0.0.7
cryptography>=42.0.0
pytest>=8.0.0
PyGithub>=2.1.1